+++ .idea/compiler.xml
@@ -30,7 +30,7 @@
           <entry name="$USER_HOME$/.gradle/caches/modules-2/files-2.1/org.jetbrains/annotations/13.0/919f0dfe192fb4e063e7dacadee7f8bb9a2672a9/annotations-13.0.jar" />
           <entry name="$USER_HOME$/.gradle/caches/modules-2/files-2.1/org.jetbrains.kotlin/kotlin-stdlib-common/1.6.10/c118700e3a33c8a0d9adc920e9dec0831171925/kotlin-stdlib-common-1.6.10.jar" />
         </processorPath>
-        <module name="julielerche-capstone.app.main" />
+        <module name="julielerche-capstone.TaskWarriorService.main" />
       </profile>
     </annotationProcessing>
     <bytecodeTargetLevel target="11" />
