+++ .idea/gradle.xml
@@ -9,7 +9,7 @@
         <option name="modules">
           <set>
             <option value="$PROJECT_DIR$" />
-            <option value="$PROJECT_DIR$/app" />
+            <option value="$PROJECT_DIR$/TaskWarriorService" />
           </set>
         </option>
       </GradleProjectSettings>
