+++ .idea/modules.xml
@@ -2,7 +2,7 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
-      <module fileurl="file://$PROJECT_DIR$/.idea/modules/app/julielerche-capstone.app.main.iml" filepath="$PROJECT_DIR$/.idea/modules/app/julielerche-capstone.app.main.iml" />
+      <module fileurl="file://$PROJECT_DIR$/.idea/modules/TaskWarriorService/julielerche-capstone.TaskWarriorService.main.iml" filepath="$PROJECT_DIR$/.idea/modules/TaskWarriorService/julielerche-capstone.TaskWarriorService.main.iml" />
     </modules>
   </component>
 </project>
