+++ TaskWarriorService/src/main/java/julielerche/capstone/dependencies/MetricsModule.java
@@ -1,0 +1,4 @@
+package julielerche.capstone.dependencies;
+
+public class MetricsModule {
+}
