--- app/src/main/java/julielerche/capstone/dependencies/MetricsModule.java
@@ -1,4 +1,0 @@
-package julielerche.capstone.dependencies;
-
-public class MetricsModule {
-}
